refactor: rename FilterControl to FilterSliderInput
